chore: move home components into home folder to consolidate
